refactor(context-menu): H.1 — extract helpers + constants
